Escape double quotes to work with spaces
It looks like I made this change a while ago and never committed it.
It looks like I am escaping double quotes around variables that would
be a path so I am guessing this was to deal with spaces in path names.
I checked the current file with install paths with spaces and it seems
to work.